refactor: reorganize assets and update styles path
